Added Display of Reviews in LikedPlaces and Fixed half star rating choice

+++ app/src/main/java/com/example/wanderlist/data/firestore/model/Reviews.kt
@@ -4,6 +4,6 @@ data class Reviews(
     val id: String = "",
     val userId: String = "",
     val establishmentId: String = "",
-    val rating: Float = 0f,
+    val rating: Int = 0,
     val reviewText: String = "",
 )
